Add Default Size as the Option
